requirements.txt
Django==1.9.5
django-registration-redux==1.4
hellowebapp-deploy
psycopg2==2.6
Pillow==3.2.0
stripe==1.32.2
djangorestframework==3.3.3


